+++ src/jk_utils/duration.py
@@ -1,5 +1,7 @@
 
 
+import typing
+import time
 
 
 
@@ -34,7 +36,64 @@ _GREGORIAN_CALENDAR_SECONDS_PER_MINUTE = 60
 # <c>(years, months, weeks, -1, -1, -1, -1)</c> is returned.
 # As this function is intended to calculate values for displaying time durations, a better accuracy doesn't make any sense anyway.
 #
-def secondsToYMWDHMS(fSeconds:float) -> tuple:
+def secondsToDHMSf(fSeconds:float) -> typing.Tuple[int,int,int,float]:
+	assert isinstance(fSeconds, (float,int))
+	assert fSeconds >= 0
+
+	if fSeconds >= _GREGORIAN_CALENDAR_SECONDS_PER_DAY:
+		days = int(fSeconds / _GREGORIAN_CALENDAR_SECONDS_PER_DAY)
+		fSeconds = fSeconds - days * _GREGORIAN_CALENDAR_SECONDS_PER_DAY
+	else:
+		days = 0
+
+	if fSeconds >= _GREGORIAN_CALENDAR_SECONDS_PER_HOUR:
+		hours = int(fSeconds / _GREGORIAN_CALENDAR_SECONDS_PER_HOUR)
+		fSeconds = fSeconds - hours * _GREGORIAN_CALENDAR_SECONDS_PER_HOUR
+	else:
+		hours = 0
+
+	if fSeconds >= _GREGORIAN_CALENDAR_SECONDS_PER_MINUTE:
+		minutes = int(fSeconds / _GREGORIAN_CALENDAR_SECONDS_PER_MINUTE)
+		fSeconds = fSeconds - minutes * _GREGORIAN_CALENDAR_SECONDS_PER_MINUTE
+	else:
+		minutes = 0
+
+	return (days, hours, minutes, fSeconds)
+#
+
+
+
+#
+# Splits the specified number of seconds into years, months, weeks, days, hours, minutes and seconds.
+#
+# As long as the time remains below the amount of one month all return values can be calculated absolutely accurately. In that case
+# <c>(0, 0, weeks, days, hours, minutes, seconds)</c> is returned.
+#
+# If the time is greater or equal to the duration of one month this can no longer be calculated accurately, as the length of the months differ. In that case 
+# <c>(years, months, weeks, -1, -1, -1, -1)</c> is returned.
+# As this function is intended to calculate values for displaying time durations, a better accuracy doesn't make any sense anyway.
+#
+def secondsToDHMS(fSeconds:float) -> typing.Tuple[int,int,int,int]:
+	days, hours, minutes, fSeconds = secondsToDHMSf(fSeconds)
+
+	seconds = int(fSeconds)
+
+	return (days, hours, minutes, seconds)
+#
+
+
+
+#
+# Splits the specified number of seconds into years, months, weeks, days, hours, minutes and seconds.
+#
+# As long as the time remains below the amount of one month all return values can be calculated absolutely accurately. In that case
+# <c>(0, 0, weeks, days, hours, minutes, seconds)</c> is returned.
+#
+# If the time is greater or equal to the duration of one month this can no longer be calculated accurately, as the length of the months differ. In that case 
+# <c>(years, months, weeks, -1, -1, -1, -1)</c> is returned.
+# As this function is intended to calculate values for displaying time durations, a better accuracy doesn't make any sense anyway.
+#
+def secondsToYMWDHMSf(fSeconds:float) -> typing.Tuple[int,int,int,int,int,int,float]:
 	assert isinstance(fSeconds, (float,int))
 	assert fSeconds >= 0
 
@@ -77,10 +136,30 @@ def secondsToYMWDHMS(fSeconds:float) -> tuple:
 	else:
 		minutes = 0
 
+	return (years, months, weeks, days, hours, minutes, fSeconds)
+#
+
+
+
+#
+# Splits the specified number of seconds into years, months, weeks, days, hours, minutes and seconds.
+#
+# As long as the time remains below the amount of one month all return values can be calculated absolutely accurately. In that case
+# <c>(0, 0, weeks, days, hours, minutes, seconds)</c> is returned.
+#
+# If the time is greater or equal to the duration of one month this can no longer be calculated accurately, as the length of the months differ. In that case 
+# <c>(years, months, weeks, -1, -1, -1, -1)</c> is returned.
+# As this function is intended to calculate values for displaying time durations, a better accuracy doesn't make any sense anyway.
+#
+def secondsToYMWDHMS(fSeconds:float) -> typing.Tuple[int,int,int,int,int,int,int]:
+	years, months, weeks, days, hours, minutes, fSeconds = secondsToYMWDHMSf(fSeconds)
+
 	seconds = int(fSeconds)
 
 	return (years, months, weeks, days, hours, minutes, seconds)
 #
+
+
 
 
 
@@ -175,6 +254,103 @@ def secondsToHRStr(fSeconds:float) -> str:
 	else:
 		return "now"
 #
+
+
+
+
+
+class DurationMeter(object):
+
+	################################################################################################################################
+	## Constants
+	################################################################################################################################
+
+	################################################################################################################################
+	## Constructor
+	################################################################################################################################
+
+	#
+	# Constructor method.
+	#
+	def __init__(self, nRepeats:int = 1, factor:typing.Union[int,float] = 1):
+		assert isinstance(nRepeats, int)
+		assert nRepeats > 0
+		assert isinstance(factor, (int, float))
+
+		self.__nRepeats = nRepeats
+		self.__factor = factor
+		self.__tStart:float = -1
+		self.__tEnd:float = -1
+		self.__tDuration:float = 0
+	#
+
+	################################################################################################################################
+	## Public Properties
+	################################################################################################################################
+
+	@property
+	def tStart(self) -> float:
+		return self.__tStart
+	#
+
+	@property
+	def tDuration(self) -> float:
+		return self.__tDuration
+	#
+
+	@property
+	def tEnd(self) -> float:
+		return self.__tEnd
+	#
+
+	@property
+	def durationStr(self, defaultValue = "not_started"):
+		if self.__tStart < 0:
+			return defaultValue
+		days, hours, minutes, fSeconds = secondsToDHMSf(self.__tDuration)
+
+		return f"{days:02d}:{hours:02d}:{minutes:02d}:{fSeconds:02f}"
+	#
+
+	################################################################################################################################
+	## Helper Methods
+	################################################################################################################################
+
+	################################################################################################################################
+	## Public Methods
+	################################################################################################################################
+
+	def __str__(self):
+		if self.__tStart < 0:
+			return "not_started"
+		return f"{self.__tDuration}"
+	#
+
+	def __repr__(self):
+		return "DurationMeter<( " + self.__str__() + ")>"
+	#
+
+	def __enter__(self):
+		self.__tEnd = -1
+		self.__tStart = time.time()
+		return self
+	#
+
+	def __exit__(self, type, value, traceback):
+		self.__tEnd = time.time()
+		self.__tDuration = (self.__tEnd - self.__tStart) * self.__factor / self.__nRepeats
+	#
+
+	################################################################################################################################
+	## Public Static Methods
+	################################################################################################################################
+
+#
+
+
+
+
+